requirements.txt
# =========================
# Web Backend
# =========================
Flask>=3.0,<4.0

# =========================
# LangChain Core
# =========================
langchain>=0.2.10,<0.3.0
langchain-core>=0.2.25,<0.3.0

# =========================
# LLM Providers
# =========================
langchain-openai>=0.1.22,<0.2.0
openai>=1.30.0,<2.0.0

# =========================
# Environment Variables
# =========================
python-dotenv>=1.0.1,<2.0.0

# =========================
# Database (MySQL)
# =========================
SQLAlchemy>=2.0.30,<3.0.0
PyMySQL>=1.1.1,<2.0.0

# =========================
# Data Handling (Excel, etc.)
# =========================
numpy>=1.26,<2.0
pandas>=2.2,<3.0
openpyxl>=3.1,<4.0

# =========================
# Dashboard
# =========================
streamlit>=1.35,<2.0
